+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# Unreleased
+
+- Add support for manipulating comments
+
 ## 0.1.0 (7 March 2026)
 
 - Initial release